feat: apply filters server-side to eliminate layout flashing
