+++ src/app/(core)/programmes/[subjectPhaseSlug]/[tab]/page.tsx
@@ -29,6 +29,10 @@ import { getMvRefreshTime } from "@/pages-helpers/curriculum/downloads/getMvRefr
 import { getFeatureFlagValue } from "@/utils/featureFlags";
 import { resolveOakHref } from "@/common-lib/urls";
 import { getSubjectPhaseSlug } from "@/components/TeacherComponents/helpers/getSubjectPhaseSlug";
+import {
+  getDefaultFilter,
+  mergeInFilterParams,
+} from "@/utils/curriculum/filtersUrl";
 
 const reportError = errorReporter("programme-page::app");
 
@@ -177,6 +181,24 @@ const InnerProgrammePage = async (props: AppPageProps<ProgrammePageParams>) => {
   const curriculumUnitsFormattedData =
     formatCurriculumUnitsData(curriculumUnitsData);
 
+  // Resolve filter server-side from URL search params
+  const defaultFilter = getDefaultFilter(curriculumUnitsFormattedData);
+  const searchParamsMap = new URLSearchParams();
+  if (searchParams) {
+    Object.entries(searchParams).forEach(([k, v]) => {
+      if (v != null) {
+        if (Array.isArray(v)) {
+          v.forEach((item) => {
+            searchParamsMap.append(k, item);
+          });
+        } else {
+          searchParamsMap.append(k, v);
+        }
+      }
+    });
+  }
+  const resolvedFilter = mergeInFilterParams(defaultFilter, searchParamsMap);
+
   // All KS4 options for subject phase
   const ks4Options =
     curriculumPhaseOptions.subjects.find(
@@ -218,6 +240,7 @@ const InnerProgrammePage = async (props: AppPageProps<ProgrammePageParams>) => {
     curriculumInfo: cachedData.programmeUnitsData,
     curriculumDownloadsTabData,
     mvRefreshTime,
+    initialFilter: resolvedFilter,
   };
 
   return <ProgrammeView {...results} />;
